requirements.txt
absl-py==0.7.1
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
bert==2.2.0
bert-serving-client==1.8.8
bert-serving-server==1.8.8
bleach==3.1.0
boto3==1.9.133
botocore==1.12.133
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.5.0
docutils==0.14
entrypoints==0.3
erlastic==2.0.0
Flask==1.0.2
gast==0.2.2
gluonnlp==0.6.0
GPUtil==1.4.0
graphviz==0.8.4
grpcio==1.19.0
h5py==2.9.0
icc-rt==2019.0
idna==2.6
intel-openmp==2019.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
MarkupSafe==1.1.1
mistune==0.8.4
mkl==2019.0
mkl-fft==1.0.6
mkl-random==1.0.1.1
mock==2.0.0
mxnet-cu92mkl==1.4.0.post0
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.7
numpy==1.16.0
olefile==0.46
opencv-contrib-python==3.3.0.10
opencv-python==3.3.0.10
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
pbr==5.1.3
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
pyext==0.7
Pygments==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytorch-pretrained-bert==0.6.1
pytz==2018.9
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.4.3
regex==2019.4.14
requests==2.18.4
s3transfer==0.2.0
Send2Trash==1.5.0
six==1.12.0
tbb==2019.0
tbb4py==2019.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-serving-api==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
torch==1.0.1
torchvision==0.2.2.post3
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.22
virtualenv==16.5.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.1
widgetsnbextension==3.4.2
